SA1402: File may only contain a single type.
Better handled better by AV1507 as it allows generic overloads in the same file
